Updating the color logic
Introduces new variables colorInt and colorCount. colorInt is colorCount divided by 255, so colorInt ranges from 0 to 1. This means that when colorCount varies from 0 to 255, colorInt corresponds to the range 0 to 1. This variable is used to make the color change softly and not drastically by stepping from 0 to 1 and then from 1 to 0 in a period of 10ms.
Also replace the random() function used in all drawing of graphics to solve the problem of color change flashing when playing music.
I tried to randomly generate pink curves at the circle's center, but the actual effect was not visually appealing, so I commented out the corresponding code. Next time it can be modified if a better presentation comes to mind.